// src/app/models/restaurant.model.ts
// Generated by https://quicktype.io

export interface RestaurantModel {
  ID: string;
  Name: string;
  Description: string;
  Address: string;
  ZipCode: string;
  Phone: string;
  OpenTime: string;
  Picture: Picture;
  Position: Position;
  Class: string;
  City: string;
  SrcUpdateTime: Date;
  UpdateTime: Date;
}



export interface Picture {
  PictureUrl1: string;
  PictureDescription1: string;
}

export interface Position {
  PositionLon: number;
  PositionLat: number;
  GeoHash: string;
}